Implement table creation with custom data types
Adds the ability to specify the data types of each column in the creation of a table using the `create_table` method of the `SqliteHandle` class. Now it's possible to specify the data types when calling the method by passing an array with the data types as the third argument.

This makes table creation more flexible and allows data types to be customized according to the needs of each table.